backend/requirements.txt
# Python dependencies for email sending
python-dotenv==1.0.0
